+++ splunk-devops/src/main/java/com/splunk/splunkjenkins/utils/RemoteUtils.java
@@ -1,6 +1,7 @@
 package com.splunk.splunkjenkins.utils;
 
 import com.splunk.splunkjenkins.SplunkJenkinsInstallation;
+import hudson.util.Secret;
 import org.apache.commons.beanutils.BeanUtils;
 
 import java.util.Map;
@@ -11,7 +12,9 @@ public class RemoteUtils {
         // Init SplunkJenkins global config in slave, can not reference Jenkins.getInstance(), Xtream
         SplunkJenkinsInstallation config = new SplunkJenkinsInstallation(false);
         try {
+            String tokenValue = (String) eventCollectorProperty.remove("token");
             BeanUtils.populate(config, eventCollectorProperty);
+            config.setToken(Secret.fromString(tokenValue));
             config.setEnabled(true);
             initSplunkConfigOnAgent(config);
         } catch (Exception e) {